Move unit tests to unittest package
and make sure they are enabled
