+++ atlasdb-cassandra-tests/build.gradle
@@ -27,7 +27,8 @@ test {
         classDumpFile = file("$buildDir/jacoco/classpathdumps")
     }
     maxHeapSize = "1024m"
-    include '**/*Suite.class'
+    include 'com/palantir/atlasdb/keyvalue/cassandra/*Suite.class'
+    include 'com/palantir/atlasdb/keyvalue/cassandra/unittest/**'
     maxParallelForks 2
     dependsOn ':atlasdb-ete-test-utils:buildCassandraImage'
 }
