+++ atlasdb-cassandra-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/unittest/CassandraJmxCompactionManagerTest.java
@@ -1,19 +1,19 @@
 /**
- * Copyright 2015 Palantir Technologies
- * <p>
+ * Copyright 2016 Palantir Technologies
+ *
  * Licensed under the BSD-3 License (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
+ *
  * http://opensource.org/licenses/BSD-3-Clause
- * <p>
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.keyvalue.cassandra;
+package com.palantir.atlasdb.keyvalue.cassandra.unittest;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
